+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Settings dialog: Beep at the selected volume when the beep volume is changed
+
 ## [0.8.0] - 2026-08-24
 
 ### Added
@@ -364,6 +370,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of web app
 
+[Unreleased]: https://github.com/treiher/valens/compare/v0.8.0...HEAD
 [0.8.0]: https://github.com/treiher/valens/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/treiher/valens/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/treiher/valens/compare/v0.5.0...v0.6.0